+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/action/ReportDefinitionActions.kt
@@ -19,9 +19,7 @@ package com.amazon.opendistroforelasticsearch.reportsscheduler.action
 import com.amazon.opendistroforelasticsearch.commons.authuser.User
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
 import com.amazon.opendistroforelasticsearch.reportsscheduler.index.ReportDefinitionsIndex
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.CreateReportDefinitionRequest
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.CreateReportDefinitionResponse
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.DeleteReportDefinitionRequest
@@ -78,12 +76,12 @@ internal object ReportDefinitionActions {
         val currentReportDefinitionDetails = ReportDefinitionsIndex.getReportDefinition(request.reportDefinitionId)
         currentReportDefinitionDetails
             ?: run {
-                MyMetrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
+                Metrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
                 throw ElasticsearchStatusException("Report Definition ${request.reportDefinitionId} not found", RestStatus.NOT_FOUND)
             }
 
         if (!UserAccessManager.doesUserHasAccess(user, currentReportDefinitionDetails.tenant, currentReportDefinitionDetails.access)) {
-            MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+            Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
             throw ElasticsearchStatusException("Permission denied for Report Definition ${request.reportDefinitionId}", RestStatus.FORBIDDEN)
         }
         val currentTime = Instant.now()
@@ -95,7 +93,7 @@ internal object ReportDefinitionActions {
             request.reportDefinition
         )
         if (!ReportDefinitionsIndex.updateReportDefinition(request.reportDefinitionId, reportDefinitionDetails)) {
-            MyMetrics.REPORT_DEFINITION_UPDATE_SYSTEM_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_UPDATE_SYSTEM_ERROR.counter.increment()
             throw ElasticsearchStatusException("Report Definition Update failed", RestStatus.INTERNAL_SERVER_ERROR)
         }
         return UpdateReportDefinitionResponse(request.reportDefinitionId)
@@ -112,12 +110,12 @@ internal object ReportDefinitionActions {
         val reportDefinitionDetails = ReportDefinitionsIndex.getReportDefinition(request.reportDefinitionId)
         reportDefinitionDetails
             ?: run {
-                MyMetrics.REPORT_DEFINITION_INFO_SYSTEM_ERROR.counter.increment()
+                Metrics.REPORT_DEFINITION_INFO_SYSTEM_ERROR.counter.increment()
                 throw ElasticsearchStatusException("Report Definition ${request.reportDefinitionId} not found", RestStatus.NOT_FOUND)
             }
 
         if (!UserAccessManager.doesUserHasAccess(user, reportDefinitionDetails.tenant, reportDefinitionDetails.access)) {
-            MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+            Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
             throw ElasticsearchStatusException("Permission denied for Report Definition ${request.reportDefinitionId}", RestStatus.FORBIDDEN)
         }
         return GetReportDefinitionResponse(reportDefinitionDetails, UserAccessManager.hasAllInfoAccess(user))
@@ -134,16 +132,16 @@ internal object ReportDefinitionActions {
         val reportDefinitionDetails = ReportDefinitionsIndex.getReportDefinition(request.reportDefinitionId)
         reportDefinitionDetails
             ?: run {
-                MyMetrics.REPORT_DEFINITION_DELETE_USER_ERROR.counter.increment()
+                Metrics.REPORT_DEFINITION_DELETE_USER_ERROR.counter.increment()
                 throw ElasticsearchStatusException("Report Definition ${request.reportDefinitionId} not found", RestStatus.NOT_FOUND)
             }
 
         if (!UserAccessManager.doesUserHasAccess(user, reportDefinitionDetails.tenant, reportDefinitionDetails.access)) {
-            MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+            Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
             throw ElasticsearchStatusException("Permission denied for Report Definition ${request.reportDefinitionId}", RestStatus.FORBIDDEN)
         }
         if (!ReportDefinitionsIndex.deleteReportDefinition(request.reportDefinitionId)) {
-            MyMetrics.REPORT_DEFINITION_DELETE_SYSTEM_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_DELETE_SYSTEM_ERROR.counter.increment()
             throw ElasticsearchStatusException("Report Definition ${request.reportDefinitionId} delete failed", RestStatus.REQUEST_TIMEOUT)
         }
         return DeleteReportDefinitionResponse(request.reportDefinitionId)
