+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/CreateReportDefinitionRequest.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.model
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_DEFINITION_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.createJsonParser
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.logger
@@ -80,7 +78,7 @@ internal class CreateReportDefinitionRequest : ActionRequest, ToXContentObject {
             }
         }
         reportDefinition ?: run {
-            MyMetrics.REPORT_DEFINITION_CREATE_USER_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_CREATE_USER_ERROR.counter.increment()
             throw IllegalArgumentException("$REPORT_DEFINITION_FIELD field absent")
         }
         this.reportDefinition = reportDefinition
