+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/CreateReportDefinitionResponse.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.model
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_DEFINITION_ID_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.logger
 import org.elasticsearch.common.io.stream.StreamInput
@@ -72,7 +70,7 @@ internal class CreateReportDefinitionResponse(
                 }
             }
             reportDefinitionId ?: run {
-                MyMetrics.REPORT_DEFINITION_CREATE_SYSTEM_ERROR.counter.increment()
+                Metrics.REPORT_DEFINITION_CREATE_SYSTEM_ERROR.counter.increment()
                 throw IllegalArgumentException("$REPORT_DEFINITION_ID_FIELD field absent")
             }
             return CreateReportDefinitionResponse(reportDefinitionId)
