+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/GetReportDefinitionResponse.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.model
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.createJsonParser
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.logger
 import org.elasticsearch.common.io.stream.StreamInput
@@ -78,7 +76,7 @@ internal class GetReportDefinitionResponse : BaseResponse {
             }
         }
         reportDefinition ?: run {
-            MyMetrics.REPORT_DEFINITION_INFO_SYSTEM_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_INFO_SYSTEM_ERROR.counter.increment()
             throw IllegalArgumentException("${RestTag.REPORT_DEFINITION_FIELD} field absent")
         }
         this.reportDefinitionDetails = reportDefinition
