+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/GetReportInstanceRequest.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.model
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_INSTANCE_ID_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.logger
 import org.elasticsearch.action.ActionRequest
@@ -77,7 +75,7 @@ internal class GetReportInstanceRequest(
                 }
             }
             reportInstanceId ?: run {
-                MyMetrics.REPORT_INSTANCE_INFO_USER_ERROR.counter.increment()
+                Metrics.REPORT_INSTANCE_INFO_USER_ERROR.counter.increment()
                 throw IllegalArgumentException("$REPORT_INSTANCE_ID_FIELD field absent")
             }
