+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/RestTag.kt
@@ -42,10 +42,6 @@ internal object RestTag {
     const val FROM_INDEX_FIELD = "fromIndex"
     const val MAX_ITEMS_FIELD = "maxItems"
     const val RETRY_AFTER_FIELD = "retryAfter"
-    const val SUCCESS_START_STATUS_CODE = 200
-    const val SUCCESS_END_STATUS_CODE = 299
-    const val CLIENT_START_STATUS_CODE = 400
-    const val CLIENT_END_STATUS_CODE = 499
     private val INCLUDE_ID = Pair(ID_FIELD, "true")
     private val EXCLUDE_ACCESS = Pair(ACCESS_LIST_FIELD, "false")
     val INSTANCE_INDEX_PARAMS: Params = ToXContent.MapParams(mapOf(INCLUDE_ID))
