+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/model/UpdateReportDefinitionRequest.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.model
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.LOG_PREFIX
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_DEFINITION_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_DEFINITION_ID_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.createJsonParser
@@ -87,11 +85,11 @@ internal class UpdateReportDefinitionRequest : ActionRequest, ToXContentObject {
             }
         }
         reportDefinitionId ?: run {
-            MyMetrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
             throw IllegalArgumentException("$REPORT_DEFINITION_ID_FIELD field absent")
         }
         reportDefinition ?: run {
-            MyMetrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
+            Metrics.REPORT_DEFINITION_UPDATE_USER_ERROR.counter.increment()
             throw IllegalArgumentException("$REPORT_DEFINITION_FIELD field absent")
         }
         this.reportDefinitionId = reportDefinitionId
