+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/resthandler/OnDemandReportRestHandler.kt
@@ -19,15 +19,12 @@ import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPl
 import com.amazon.opendistroforelasticsearch.reportsscheduler.action.InContextReportCreateAction
 import com.amazon.opendistroforelasticsearch.reportsscheduler.action.OnDemandReportCreateAction
 import com.amazon.opendistroforelasticsearch.reportsscheduler.action.ReportInstanceActions
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.InContextReportCreateRequest
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.OnDemandReportCreateRequest
 import com.amazon.opendistroforelasticsearch.reportsscheduler.model.RestTag.REPORT_DEFINITION_ID_FIELD
 import com.amazon.opendistroforelasticsearch.reportsscheduler.util.contentParserNextToken
 import org.elasticsearch.client.node.NodeClient
-import org.elasticsearch.rest.BaseRestHandler
 import org.elasticsearch.rest.BaseRestHandler.RestChannelConsumer
 import org.elasticsearch.rest.BytesRestResponse
 import org.elasticsearch.rest.RestHandler.Route
@@ -40,7 +37,7 @@ import org.elasticsearch.rest.RestStatus
  * Rest handler for creating on-demand report instances.
  * This handler uses [ReportInstanceActions].
  */
-internal class OnDemandReportRestHandler : BaseRestHandler() {
+internal class OnDemandReportRestHandler : PluginBaseHandler() {
     companion object {
         private const val REPORT_INSTANCE_LIST_ACTION = "on_demand_report_actions"
         private const val ON_DEMAND_REPORT_URL = "$BASE_REPORTS_URI/on_demand"
@@ -86,18 +83,18 @@ internal class OnDemandReportRestHandler : BaseRestHandler() {
     /**
      * {@inheritDoc}
      */
-    override fun prepareRequest(request: RestRequest, client: NodeClient): RestChannelConsumer {
+    override fun executeRequest(request: RestRequest, client: NodeClient): RestChannelConsumer {
         return when (request.method()) {
             PUT -> RestChannelConsumer {
-                MyMetrics.REPORT_FROM_DEFINITION_TOTAL.counter.increment()
-                MyMetrics.REPORT_FROM_DEFINITION_INTERVAL_COUNT.counter.increment()
+                Metrics.REPORT_FROM_DEFINITION_TOTAL.counter.increment()
+                Metrics.REPORT_FROM_DEFINITION_INTERVAL_COUNT.counter.increment()
                 client.execute(InContextReportCreateAction.ACTION_TYPE,
                     InContextReportCreateRequest(request.contentParserNextToken()),
                     RestResponseToXContentListener(it))
             }
             POST -> RestChannelConsumer {
-                MyMetrics.REPORT_FROM_DEFINITION_ID_TOTAL.counter.increment()
-                MyMetrics.REPORT_FROM_DEFINITION_ID_INTERVAL_COUNT.counter.increment()
+                Metrics.REPORT_FROM_DEFINITION_ID_TOTAL.counter.increment()
+                Metrics.REPORT_FROM_DEFINITION_ID_INTERVAL_COUNT.counter.increment()
                 client.execute(OnDemandReportCreateAction.ACTION_TYPE,
                     OnDemandReportCreateRequest.parse(request.contentParserNextToken(), request.param(REPORT_DEFINITION_ID_FIELD)),
                     RestResponseToXContentListener(it))
