+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/resthandler/ReportStatsRestHandler.kt
@@ -16,7 +16,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.resthandler
 
 import com.amazon.opendistroforelasticsearch.reportsscheduler.ReportsSchedulerPlugin.Companion.BASE_REPORTS_URI
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import org.elasticsearch.client.node.NodeClient
 import org.elasticsearch.rest.BaseRestHandler
 import org.elasticsearch.rest.BaseRestHandler.RestChannelConsumer
@@ -50,7 +50,7 @@ internal class ReportStatsRestHandler : BaseRestHandler() {
             /**
              * Get reporting backend stats
              * Request URL: GET REPORT_STATS_URL
-             * Response body derived from: Ref [com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName]
+             * Response body derived from: Ref [com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics]
              */
             Route(GET, "$REPORT_STATS_URL")
         )
@@ -68,10 +68,10 @@ internal class ReportStatsRestHandler : BaseRestHandler() {
      */
     override fun prepareRequest(request: RestRequest, client: NodeClient): RestChannelConsumer {
         return when (request.method()) {
-            // TODO: Wrap this API in TransportAction
+            // TODO: Wrap this into TransportAction
             GET -> RestChannelConsumer {
                 // it.sendResponse(BytesRestResponse(RestStatus.OK, Metrics.getInstance().collectToFlattenedJSON()))
-                it.sendResponse(BytesRestResponse(RestStatus.OK, MyMetrics.collectToFlattenedJSON()))
+                it.sendResponse(BytesRestResponse(RestStatus.OK, Metrics.collectToFlattenedJSON()))
             }
             else -> RestChannelConsumer {
                 it.sendResponse(BytesRestResponse(RestStatus.METHOD_NOT_ALLOWED, "${request.method()} is not allowed"))
