+++ reports-scheduler/src/main/kotlin/com/amazon/opendistroforelasticsearch/reportsscheduler/security/UserAccessManager.kt
@@ -17,9 +17,7 @@
 package com.amazon.opendistroforelasticsearch.reportsscheduler.security
 
 import com.amazon.opendistroforelasticsearch.commons.authuser.User
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
-// import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MetricName
-import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.MyMetrics
+import com.amazon.opendistroforelasticsearch.reportsscheduler.metrics.Metrics
 import com.amazon.opendistroforelasticsearch.reportsscheduler.settings.PluginSettings
 import com.amazon.opendistroforelasticsearch.reportsscheduler.settings.PluginSettings.FilterBy
 import org.elasticsearch.ElasticsearchStatusException
@@ -51,7 +49,7 @@ internal object UserAccessManager {
      */
     fun validateUser(user: User?) {
         if (isUserPrivateTenant(user) && user?.name == null) {
-            MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+            Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
             throw ElasticsearchStatusException("User name not provided for private tenant access",
                 RestStatus.FORBIDDEN)
         }
@@ -61,25 +59,25 @@ internal object UserAccessManager {
             FilterBy.User -> { // User name must be present
                 user?.name
                     ?: run {
-                        MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+                        Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
                         throw ElasticsearchStatusException("Filter-by enabled with security disabled",
                             RestStatus.FORBIDDEN)
                     }
             }
             FilterBy.Roles -> { // backend roles must be present
                 if (user == null || user.roles.isNullOrEmpty()) {
-                    MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+                    Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
                     throw ElasticsearchStatusException("User doesn't have roles configured. Contact administrator.",
                         RestStatus.FORBIDDEN)
                 } else if (user.roles.stream().filter { !PluginSettings.ignoredRoles.contains(it) }.count() == 0L) {
-                    MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+                    Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
                     throw ElasticsearchStatusException("No distinguishing roles configured. Contact administrator.",
                         RestStatus.FORBIDDEN)
                 }
             }
             FilterBy.BackendRoles -> { // backend roles must be present
                 if (user?.backendRoles.isNullOrEmpty()) {
-                    MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+                    Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
                     throw ElasticsearchStatusException("User doesn't have backend roles configured. Contact administrator.",
                         RestStatus.FORBIDDEN)
                 }
@@ -93,7 +91,7 @@ internal object UserAccessManager {
     fun validatePollingUser(user: User?) {
         if (user != null) { // Check only if security is enabled
             if (user.name != KIBANA_SERVER_USER) {
-                MyMetrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
+                Metrics.REPORT_PERMISSION_USER_ERROR.counter.increment()
                 throw ElasticsearchStatusException("Permission denied", RestStatus.FORBIDDEN)
             }
         }
